feat(import): move geom_from_area_code to synthese imports
