+++ backend/geonature/core/imports/checks/sql/geo.py
@@ -9,66 +9,37 @@ from geonature.utils.env import db
 
 from geonature.core.imports.checks.sql.utils import report_erroneous_rows
 
-from ref_geo.models import LAreas, BibAreasTypes
+from ref_geo.models import LAreas
 
 
 __all__ = [
-    "set_geom_from_area_code",
+    "set_geom_point",
     "convert_geom_columns",
     "check_is_valid_geography",
+    "check_geometry_defined",
     "check_geography_outside",
 ]
 
 
-def set_geom_from_area_code(
-    imprt, entity, geom_4326_col, geom_local_col, source_column, area_type_filter
-):  # XXX specific synthese
+def set_geom_point(imprt, entity, geom_4326_field, geom_point_field):
     transient_table = imprt.destination.get_transient_table()
-    # Find area in CTE, then update corresponding column in statement
-    cte = (
-        select(
-            transient_table.c.id_import,
-            transient_table.c.line_no,
-            LAreas.id_area,
-            LAreas.geom,
-            # TODO: add LAreas.geom_4326
-        )
-        .select_from(
-            join(transient_table, LAreas, source_column == LAreas.area_code).join(BibAreasTypes)
-        )
-        .where(transient_table.c.id_import == imprt.id_import)
-        .where(transient_table.c[entity.validity_column] == True)
-        .where(transient_table.c[geom_4326_col] == None)  # geom_4326 & local should be aligned
-        .where(area_type_filter)
-        .cte("cte")
-    )
+    # Set the_geom_point:
     stmt = (
         update(transient_table)
+        .where(transient_table.c.id_import == imprt.id_import)
+        .where(transient_table.c[entity.validity_column] == True)
         .values(
             {
-                transient_table.c.id_area_attachment: cte.c.id_area,
-                transient_table.c[geom_local_col]: cte.c.geom,
-                transient_table.c[geom_4326_col]: ST_Transform(
-                    cte.c.geom, 4326
-                ),  # TODO: replace with cte.c.geom_4326
+                geom_point_field.dest_field: ST_Centroid(
+                    transient_table.c[geom_4326_field.dest_field]
+                ),
             }
         )
-        .where(transient_table.c.id_import == cte.c.id_import)
-        .where(transient_table.c.line_no == cte.c.line_no)
     )
     db.session.execute(stmt)
 
 
-def convert_geom_columns(
-    imprt,
-    entity,
-    geom_4326_field,
-    geom_local_field,
-    geom_point_field=None,
-    codecommune_field=None,
-    codemaille_field=None,
-    codedepartement_field=None,
-):
+def convert_geom_columns(imprt, entity, geom_4326_field, geom_local_field):
     local_srid = db.session.execute(sa.func.Find_SRID("ref_geo", "l_areas", "geom")).scalar()
     transient_table = imprt.destination.get_transient_table()
     if geom_4326_field is None:
@@ -94,35 +65,9 @@ def convert_geom_columns(
     )
     db.session.execute(stmt)
 
-    for field, area_type_filter in [
-        (codecommune_field, BibAreasTypes.type_code == "COM"),
-        (codedepartement_field, BibAreasTypes.type_code == "DEP"),
-        (codemaille_field, BibAreasTypes.type_code.in_(["M1", "M5", "M10"])),
-    ]:
-        if field is None:
-            continue
-        source_column = transient_table.c[field.source_field]
-        # Set geom from area of the given type and with matching area_code:
-        set_geom_from_area_code(
-            imprt,
-            entity,
-            geom_4326_field.dest_field,
-            geom_local_field.dest_field,
-            source_column,
-            area_type_filter,
-        )
-        # Mark rows with code specified but geom still empty as invalid:
-        report_erroneous_rows(
-            imprt,
-            entity,
-            error_type="INVALID_ATTACHMENT_CODE",
-            error_column=field.name_field,
-            whereclause=sa.and_(
-                transient_table.c[geom_4326_field.dest_field] == None,
-                transient_table.c[entity.validity_column] == True,
-                source_column != None,
-            ),
-        )
+
+def check_geometry_defined(imprt, entity, geom_4326_field):
+    transient_table = imprt.destination.get_transient_table()
     # Mark rows with no geometry as invalid:
     report_erroneous_rows(
         imprt,
@@ -134,21 +79,6 @@ def convert_geom_columns(
             transient_table.c[entity.validity_column] == True,
         ),
     )
-    # Set the_geom_point:
-    if geom_point_field is not None:
-        stmt = (
-            update(transient_table)
-            .where(transient_table.c.id_import == imprt.id_import)
-            .where(transient_table.c[entity.validity_column] == True)
-            .values(
-                {
-                    geom_point_field.dest_field: ST_Centroid(
-                        transient_table.c[geom_4326_field.dest_field]
-                    ),
-                }
-            )
-        )
-        db.session.execute(stmt)
 
 
 def check_is_valid_geography(imprt, entity, wkt_field, geom_field):
